requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
fastapi==0.110.0
h11==0.14.0
idna==3.6
kombu==5.3.5
pillow==10.2.0
prompt-toolkit==3.0.43
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.3
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.19.0
vine==5.1.0
wcwidth==0.2.13
